+++ Arale.CodeGen/Arale.CodeGen.Commons/Constants/TemplateName.cs
@@ -51,5 +51,17 @@ public enum TemplateName
     /// <summary>
     ///     Hibernate (JPA) entity template
     /// </summary>
-    [Description("Java/HibernateEntity")] HibernateEntity
+    [Description("Java/HibernateEntity")] HibernateEntity,
+
+    /// <summary>
+    ///     Spring Data MongoDB entity template
+    /// </summary>
+    [Description("Java/SpringDataMongoDBEntity")]
+    SpringDataMongoDbEntity,
+
+    /// <summary>
+    ///     CSharp MongoDB.Driver entity template
+    /// </summary>
+    [Description("CSharp/CSharpMongoDBDriverEntity")]
+    CSharpMongoDbDriverEntity
 }