+++ Arale.CodeGen/Arale.CodeGen.Infrastructure/Generators/CSharpMongoDbDriverEntityCodeGenerator.cs
@@ -1,0 +1,14 @@
+﻿using Arale.CodeGen.Commons.Constants;
+
+namespace Arale.CodeGen.Infrastructure.Generators;
+
+/// <summary>
+/// </summary>
+public class CSharpMongoDbDriverEntityCodeGenerator : BaseCodeGenerator
+{
+    /// <inheritdoc />
+    protected override TemplateName GetTemplateName()
+    {
+        return TemplateName.CSharpMongoDbDriverEntity;
+    }
+}