+++ Arale.CodeGen/Arale.CodeGen.Tests/JsonParserTests.cs
@@ -3,6 +3,7 @@ using Arale.CodeGen.Commons.Constants;
 using Arale.CodeGen.Infrastructure.Parsers;
 using Arale.CodeGen.Models;
 using Arale.CodeGen.Models.Dto;
+using Arale.CodeGen.Tests.Constants;
 
 namespace Arale.CodeGen.Tests;
 
@@ -14,80 +15,8 @@ public class JsonParserTests
     [Fact]
     public void TestParse()
     {
-        const string jsonStr = """
-                               [
-                                 {
-                                   "id": 1,
-                                   "username": "shiloh",
-                                   "gender": 1,
-                                   "birthday": "1998-03-02",
-                                   "email": null,
-                                   "pi": 3.1415926,
-                                   "address": {
-                                     "country": "China",
-                                     "province": "Guangdong",
-                                     "city": "FoShan",
-                                     "street": "桂城街道"
-                                   },
-                                   "doubleNumArr": [
-                                     1.52,
-                                     2.53,
-                                     3.53,
-                                     4.54,
-                                     5.55
-                                   ],
-                                   "intArr": [
-                                     1,
-                                     2,
-                                     3,
-                                     4,
-                                     5
-                                   ],
-                                   "strArr": [
-                                     "a",
-                                     "b",
-                                     "c",
-                                     "d",
-                                     "e"
-                                   ],
-                                   "roles": [
-                                     {
-                                       "id": 1,
-                                       "roleName": "admin",
-                                       "permissions": [
-                                         {
-                                           "id": 1,
-                                           "code": "sys:user:view",
-                                           "description": "查看用户列表",
-                                           "url": "/api/users"
-                                         }
-                                       ]
-                                     },
-                                     {
-                                       "id": 2,
-                                       "roleName": "user"
-                                     }
-                                   ],
-                                   "enabled": true
-                                 },
-                                 {
-                                   "id": 2,
-                                   "username": "bruce",
-                                   "gender": 1,
-                                   "birthday": "1998-03-02",
-                                   "email": "bruce@gmail.com",
-                                   "pi": 3.1415926,
-                                   "address": {
-                                     "country": "China",
-                                     "province": "Guangdong",
-                                     "city": "FoShan",
-                                     "street": "桂城街道"
-                                   },
-                                   "enabled": false
-                                 }
-                               ]
-                               """;
-        var jsonParser = new JsonParser(new CodeGenerateReq { Code = jsonStr, TargetType = TargetType.CSharpClass })
+        var jsonParser = new JsonParser(new CodeGenerateReq
+            { Code = TestDataConstants.TestJsonStr, TargetType = TargetType.CSharpClass })
         {
             RootModel = new ModelInfo
             {
@@ -96,7 +25,7 @@ public class JsonParserTests
                 Comment = ModelInfo.DefaultClassName
             }
         };
-        var jsonObj = JsonNode.Parse(jsonStr) switch
+        var jsonObj = JsonNode.Parse(TestDataConstants.TestJsonStr) switch
         {
             JsonObject jsonObject => jsonObject,
             JsonArray jsonArray => jsonArray[0]!.AsObject(),